Finished up the Dto's and rearrange the file structure
