+++ SRC/Core/LABMS.Application/DTOs/ForCreation/AmazonBooksForCreation.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LABMS.Application.DTOs
+namespace LABMS.Application.DTOs.ForCreation
 {
     public class AmazonBooksForCreation
     {
