+++ SRC/Core/LABMS.Application/DTOs/ForCreation/AuthorForCreation.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LABMS.Application.DTOs
+namespace LABMS.Application.DTOs.ForCreation
 {
     public class AuthorForCreation
     {
