+++ SRC/Core/LABMS.Application/DTOs/ForCreation/Books_At_LibraryForCreation.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LABMS.Application.DTOs.ForCreation
+{
+    public record Books_At_LibraryForCreation
+    {
+        public int LibraryId { get; init; }
+
+        public int Isbn { get; init; }
+
+        [Range(0, int.MaxValue, ErrorMessage = "Quantity cannot be negative.")]
+        public int Quantity_In_Stock { get; init; }
+    }
+}
