+++ SRC/Core/LABMS.Application/DTOs/ForCreation/LibraryForCreation.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LABMS.Application.DTOs
+namespace LABMS.Application.DTOs.ForCreation
 {
     public class LibraryForCreation
     {
