+++ SRC/Core/LABMS.Application/DTOs/ForDto/AddressDto.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LABMS.Application.DTOs
+namespace LABMS.Application.DTOs.ForDto
 {
     public class AddressDto
     {
