+++ SRC/Core/LABMS.Application/DTOs/ForDto/AmazonBooksDto.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LABMS.Application.DTOs
+namespace LABMS.Application.DTOs.ForDto
 {
     public class AmazonBooksDto
     {
