+++ SRC/Core/LABMS.Application/DTOs/ForDto/BookDto.cs
@@ -6,11 +6,14 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
 
-namespace LABMS.Application.DTOs
+namespace LABMS.Application.DTOs.ForDto
 {
-    public class BookForUpdate
+    public class BookDto
     {
-        [Key]
         public int Isbn { get; set; }
+
+        public string? BookTitle { get; set; }
+
+        public DateTime? Date_Of_Publication { get; set; }
     }
 }
