+++ SRC/Core/LABMS.Application/DTOs/ForDto/Books_At_LibraryDto.cs
@@ -1,0 +1,21 @@
+﻿using LABMS.Domain.entities;
+using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LABMS.Application.DTOs.ForDto
+{
+    public record Books_At_LibraryDto
+    {
+        public int LibraryId { get; init; }
+
+        public int Isbn { get; init; }
+
+        [Range(0, int.MaxValue, ErrorMessage = "Quantity cannot be negative.")]
+        public int Quantity_In_Stock { get; init; }
+    }
+}
