+++ SRC/Core/LABMS.Application/DTOs/ForDto/Books_By_AuthorDto.cs
@@ -1,0 +1,18 @@
+﻿using LABMS.Domain.entities;
+using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LABMS.Application.DTOs.ForDto
+{
+    public record Books_By_AuthorDto
+    {
+        public int AuthorId { get; init; }
+
+        public int Isbn { get; init; }
+    }
+}
