+++ SRC/Core/LABMS.Application/DTOs/ForDto/LibraryDto.cs
@@ -7,17 +7,16 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LABMS.Application.DTOs
+namespace LABMS.Application.DTOs.ForDto
 {
-    public class LibraryForUpdate
+    public class LibraryDto
     {
-        [Key]
         public int LibraryId { get; set; }
 
-        [ForeignKey(nameof(Address))]
         public int AddressId { get; set; }
 
-        [StringLength(200, ErrorMessage = "Library Details cannot be more than 200 words")]
+        public string? LibraryName { get; set; }
+
         public string? LibraryDetails { get; set; }
     }
 }
