+++ SRC/Core/LABMS.Application/DTOs/ForDto/MemberDto.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LABMS.Application.DTOs
+namespace LABMS.Application.DTOs.ForDto
 {
     public class MemberDto
     {
