+++ SRC/Core/LABMS.Application/DTOs/ForDto/MemberRequestDto.cs
@@ -7,20 +7,18 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LABMS.Application.DTOs
+namespace LABMS.Application.DTOs.ForDto
 {
-    public class MemberRequestForCreation
+    public class MemberRequestDto
     {
-        [ForeignKey(nameof(Member))]
+        public int RequestId { get; set; }
+
         public int MemberId { get; set; }
 
-        [ForeignKey(nameof(Book))]
         public int Isbn { get; set; }
 
-        [Required(ErrorMessage = "The date requested is required.")]
         public DateTime DateRequested { get; set; }
 
-        [Required(ErrorMessage = "The date located is required.")]
         public DateTime DateLocated { get; set; }
     }
 }
